pyproject.toml
[project]
name = "cognigraph"
version = "0.1.0"
description = "LLM-agnostic learning assistant with LangGraph, Streamlit, and Obsidian notes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph",
    "langchain",
    "langchain-openai",
    "langchain-community",
    "langchain-ollama",
    "tavily-python",
    "streamlit",
    "python-dotenv",
    "sqlalchemy",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cognigraph"]
